+++ docs/CHANGELOG.md
@@ -39,6 +39,9 @@ Version 1.9.0 (not released yet)
 - Add Kill Reward settings for dedicated servers
 - Do not load unnecessary VPPs in dedicated server mode
 - Add level filename to "Level Initializing" console message
+- Add `map_rand` command to change to a random level on the rotation
+- Add `vote rand` vote type to initiate a level change to a random level on the rotation
+- Add `$DF Dynamic Rotation` option in dedicated server config
 - Properly handle WM_PAINT in dedicated server, may improve performance (DF bug)
 - Disable adding faces to fix PS2 tiling in editor
 - Disable red background in log window if geometry limits are crossed in editor